move fetchResults function to SearchForm
Previously, fetchResults would only be trigged when App was updated. However, it makes more sense for fetchResults to run after the SearchForm updated. This will avoid any confusion in case the App component becomes more complicated in the future.

The SearchForm's input values are only relevant to that component as well, so it makes sense to keep those inputs within the component and only pass the results up to App.js.